backend/requirements.txt
django>=5.0,<7.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
drf-spectacular>=0.27
Pillow>=10.0
requests>=2.31
